1-sem/RAS/Trabalho/Parte3/PictuRAS-main/backend/ws/requirements.txt
aio_pika==9.5.4
pika==1.3.2
python-dotenv==1.0.1
Requests==2.32.3
websockets==14.1
